Hotfix: always use int for encoding

+++ bdranalytics/pipeline/encoders.py
@@ -133,7 +133,7 @@ class StringIndexer(BaseEstimator, TransformerMixin):
         for col in self.columns:
             dictionary = self.dictionaries[col]
             na_value = len(dictionary) + 1
-            transformed_column = X[col].apply(lambda x: dictionary.get(x, na_value))
+            transformed_column = X[col].apply(lambda x: dictionary.get(x, na_value)).astype(int)
             column_array.append(transformed_column.values.reshape(-1, 1))
         return np.hstack(column_array)
